+++ CHANGELOG.md
@@ -12,7 +12,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## [Unreleased]
 ### Added
+- Width of Gantt chart is now calculated on browser window width, making it full screen
 - work in progress on [clorichel/ganttlab issues](https://gitlab.com/clorichel/ganttlab/issues?scope=all&state=opened&utf8=%E2%9C%93&label_name%5B%5D=Feature)...
+
+### Fixed
+- A padding problem causing useless horizontal scrolling on the page
 
 ## 0.2.0 - 2016-11-05
 ### Added