+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - Width of Gantt chart is now calculated on browser window width, making it full screen
 - Improved and refined global UI
 - A button to refresh current display, keeping filtering and pagination state
+- One can now share or bookmark a GanttLab Live link with full issue filtering configuration
 - work in progress on [ganttlab/ganttlab-live issues](https://gitlab.com/ganttlab/ganttlab-live/issues?scope=all&state=opened&utf8=%E2%9C%93&label_name%5B%5D=Feature)...
 
 ### Changed