+++ CHANGELOG.md
@@ -12,13 +12,16 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## [Unreleased]
 ### Added
+- work in progress on [ganttlab/ganttlab-live issues](https://gitlab.com/ganttlab/ganttlab-live/issues?scope=all&state=opened&utf8=%E2%9C%93&label_name%5B%5D=Feature)...
+
+## 0.3.0 - 2016-11-24
+### Added
 - Width of Gantt chart is now calculated on browser window width, making it full screen
 - Improved and refined global UI
 - A button to refresh current display, keeping filtering and pagination state
 - One can now share or bookmark a GanttLab Live link with full issue filtering configuration
 - The expected "Remember me" checkbox, backed by browser localStorage
 - User is now able to search for projects, or for groups and projects in this group
-- work in progress on [ganttlab/ganttlab-live issues](https://gitlab.com/ganttlab/ganttlab-live/issues?scope=all&state=opened&utf8=%E2%9C%93&label_name%5B%5D=Feature)...
 
 ### Changed
 - Today vertical axis on the Gantt chart is now dashed for readability
@@ -81,4 +84,4 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - README includes a preview, and describes main topics to get started
 - Initial vue-cli scaffolding with webpack plugin
 
-[Unreleased]: https://gitlab.com/ganttlab/ganttlab-live/compare/v0.2.0...master
+[Unreleased]: https://gitlab.com/ganttlab/ganttlab-live/compare/v0.3.0...master